requirements.txt
Flask==2.1.0
mysql_connector_repackaged==0.3.1
numpy==1.24.3
opencv_python==4.7.0.72
Pillow==10.0.0
Requests==2.31.0
setuptools==66.0.0
torch==2.0.1
ultralytics==8.0.141
vietocr==0.3.11
